refactor(api_summary): isolate package:meta traits into dedicated library
