+++ pkgs/api_summary/api.txt
@@ -210,6 +210,7 @@ package:api_summary/api_summary.dart:
     isUseResult (getter: bool)
     isVisibleForOverriding (getter: bool)
     isVisibleForTesting (getter: bool)
+    hasMeta (method: bool Function(MetaContract))
 dart:async:
   Future (referenced)
 dart:core:
