+++ pkgs/api_summary/lib/api_summary.dart
@@ -17,6 +17,7 @@ export 'src/api_summary_customizer.dart'
     show ApiSummaryContext, ApiSummaryCustomizer;
 export 'src/api_trait.dart';
 export 'src/api_type.dart';
+export 'src/meta_trait.dart';
 
 /// Creates a canonical [ApiSummary] model of the public API of a package.
 ///
